+++ examples/example-cpp-app/proj.android/app/src/org/cocos2dx/cpp/AppActivity.java
@@ -30,7 +30,9 @@ import android.os.Build;
 import android.view.WindowManager;
 import android.view.WindowManager.LayoutParams;
 import com.bugsnag.android.Bugsnag;
+import com.bugsnag.android.Configuration;
 import com.bugsnag.android.BugsnagCocos2dxPlugin;
+import com.bugsnag.android.ThreadSendPolicy;
 
 public class AppActivity extends Cocos2dxActivity {
 
@@ -38,7 +40,6 @@ public class AppActivity extends Cocos2dxActivity {
     protected void onCreate(Bundle savedInstanceState) {
         super.setEnableVirtualButton(false);
         super.onCreate(savedInstanceState);
-
         // Workaround in https://stackoverflow.com/questions/16283079/re-launch-of-activity-on-home-button-but-only-the-first-time/16447508
         if (!isTaskRoot()) {
             // Android launched another instance of the root activity into an existing task
@@ -47,8 +48,16 @@ public class AppActivity extends Cocos2dxActivity {
             // Don't need to finish it again since it's finished in super.onCreate .
             return;
         }
-        BugsnagCocos2dxPlugin.register();
-        Bugsnag.init(this);
+
+        Configuration config = Configuration.load(this);
+        config.addPlugin(new BugsnagCocos2dxPlugin());
+
+        // Example custom configuration
+        config.setContext("context a");
+        config.addFeatureFlag("feature a");
+        config.setSendThreads(ThreadSendPolicy.ALWAYS);
+
+        Bugsnag.start(this, config);
 
         // Make sure we're running on Pie or higher to change cutout mode
         if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.P) {