+++ scripts/export-package.sh
@@ -18,37 +18,43 @@
 # │       ├── bugsnag-plugin-android-ndk.aar
 # │       └── bugsnag-plugin-android-anr.aar
 # ├── cocoa
-# │   ├── libbugsnag-cocos2dx-ios.a
-# │   └── libbugsnag-cocos2dx-macosx.a
+# │   └── BugsnagCocos2dx.xcframework
 # └── include
 #     └── BugsnagCocos2dx
-#         ├── cocoa/{bugsnag-cocoa public headers}
-#         ├── cocoa/BugsnagCocos2dxPlugin.h
 #         └── Bugsnag.hpp
 
 PROJ_DIR="$(cd "$(dirname "${BASH_SOURCE[0]}")"/.. && pwd)"
+BUILD_DIR="$PROJ_DIR/build"
+PKG_DIR="$BUILD_DIR/bugsnag"
+ZIPFILE="$BUILD_DIR/bugsnag-cocos2dx.zip"
 
 # Stop the script when things go wrong
-set -e pipefail
+set -xue pipefail
 
-if [ -d "$PROJ_DIR"/build/pkg ]; then
+if [ -d "$BUILD_DIR" ]; then
     echo "Deleting old artifacts"
-    rm -rf "$PROJ_DIR"/build/pkg
+    rm -rf "$BUILD_DIR"
 fi
+rm -f "$ZIPFILE"
 
 echo "Building staging directory..."
-mkdir -p "$PROJ_DIR"/build/pkg/{android,cocoa,include/BugsnagCocos2dx/cocoa}
-mkdir -p "$PROJ_DIR"/build/pkg/android/{private,libs}
+mkdir -p "$PKG_DIR"/{android,cocoa,include/BugsnagCocos2dx}
+mkdir -p "$PKG_DIR"/android/{private,libs}
 
 # Copy in headers that users will need
 install "$PROJ_DIR"/src/Bugsnag.hpp \
-    "$PROJ_DIR"/build/pkg/include/BugsnagCocos2dx
+    "$PKG_DIR"/include/BugsnagCocos2dx
 
 # Copy in CMakeLists.txt
-install "$PROJ_DIR"/src/CMakeLists.txt "$PROJ_DIR"/build/pkg/
+install "$PROJ_DIR"/src/CMakeLists.txt "$PKG_DIR"/
+
+
+###########
+# Android #
+###########
 
 echo "Assembling android artifacts..."
-mkdir -p "$PROJ_DIR"/build/android
+mkdir -p "$BUILD_DIR"/android
 pushd src/android/bugsnag-android
 GRADLE="./gradlew --quiet -PABI_FILTERS=x86,x86_64,armeabi-v7a,arm64-v8a"
     $GRADLE bugsnag-android-core:assembleRelease
@@ -62,22 +68,27 @@ ANDROID_ARTIFACTS=$(ls "$PROJ_DIR"/src/android/bugsnag-android/bugsnag-*/build/o
 # Copy in android artifacts
 for file in $ANDROID_ARTIFACTS
 do
-    cp "$file" "$PROJ_DIR"/build/pkg/android/libs/$(basename "${file/-release/}")
+    cp "$file" "$PKG_DIR"/android/libs/$(basename "${file/-release/}")
 done
 
 # Copy in Android source file
-install "$PROJ_DIR"/src/android/Bugsnag.cpp "$PROJ_DIR"/build/pkg/android
+install "$PROJ_DIR"/src/android/Bugsnag.cpp "$PKG_DIR"/android
 
 # Copy in Android private headers
-install "$PROJ_DIR"/src/android/bugsnag-android/bugsnag-plugin-android-ndk/src/main/assets/include/{bugsnag,report}.h \
-    "$PROJ_DIR"/build/pkg/android/private
+install "$PROJ_DIR"/src/android/bugsnag-android/bugsnag-plugin-android-ndk/src/main/assets/include/{bugsnag,event}.h \
+    "$PKG_DIR"/android/private
 
 # Copy in Android plugin sources
-mkdir -p "$PROJ_DIR"/build/pkg/android/bugsnag-plugin-android-cocos2dx
+mkdir -p "$PKG_DIR"/android/bugsnag-plugin-android-cocos2dx
 cp -r "$PROJ_DIR"/src/android/bugsnag-plugin-android-cocos2dx/src \
-    "$PROJ_DIR"/build/pkg/android/bugsnag-plugin-android-cocos2dx/src
+    "$PKG_DIR"/android/bugsnag-plugin-android-cocos2dx/src
 install "$PROJ_DIR"/src/android/bugsnag-plugin-android-cocos2dx/build.gradle \
-    "$PROJ_DIR"/build/pkg/android/bugsnag-plugin-android-cocos2dx
+    "$PKG_DIR"/android/bugsnag-plugin-android-cocos2dx
+
+
+#########
+# Cocoa #
+#########
 
 echo "Assembling cocoa artifacts..."
 
@@ -91,23 +102,34 @@ do
         -sdk $sdk build
 done
 
-# Join iOS artifacts into a single library
-lipo -create "$PROJ_DIR"/build/iphone*/Build/Products/Release/libBugsnagCocos2dx.a \
-    -output "$PROJ_DIR"/build/pkg/cocoa/libbugsnag-cocos2dx-ios.a
-
-cp "$PROJ_DIR"/build/macosx/Build/Products/Release/libBugsnagCocos2dx.a \
-    "$PROJ_DIR"/build/pkg/cocoa/libbugsnag-cocos2dx-macosx.a
-
-# Copy the cocoa-specific headers in
+# Get the cocoa-specific headers
+rm -rf "$BUILD_DIR"/include
+mkdir -p "$BUILD_DIR"/include/{Bugsnag,BugsnagCocos2dx}
 install $(cat "$PROJ_DIR"/src/cocoa/bugsnag-cocoa/Bugsnag.podspec.json \
     | jq .public_header_files[] \
     | xargs -n 1 -I@ bash -c "ls \"$PROJ_DIR\"/src/cocoa/bugsnag-cocoa/@") \
-    "$PROJ_DIR"/build/pkg/include/BugsnagCocos2dx/cocoa
+    "$BUILD_DIR"/include/Bugsnag
 install "$PROJ_DIR"/src/cocoa/BugsnagCocos2dxPlugin.h \
-    "$PROJ_DIR"/build/pkg/include/BugsnagCocos2dx/cocoa
+    "$BUILD_DIR"/include/BugsnagCocos2dx
+install "$PROJ_DIR"/src/Bugsnag.hpp \
+    "$BUILD_DIR"/include/BugsnagCocos2dx
+
+# Join artifacts into a single library
+xcodebuild -create-xcframework \
+    -library "$BUILD_DIR"/iphoneos/Build/Products/Release-iphoneos/libBugsnagCocos2dx.a \
+    -headers "$BUILD_DIR"/include \
+    -library "$BUILD_DIR"/iphonesimulator/Build/Products/Release-iphonesimulator/libBugsnagCocos2dx.a \
+    -headers "$BUILD_DIR"/include \
+    -library "$BUILD_DIR"/macosx/Build/Products/Release/libBugsnagCocos2dx.a \
+    -headers "$BUILD_DIR"/include \
+    -output "$PKG_DIR"/cocoa/BugsnagCocos2dx.xcframework
+
+
+###########
+# Zipfile #
+###########
 
 # Zip everything up
-pushd "$PROJ_DIR"/build/pkg
-    zip -r ../bugsnag-cocos2dx.zip .
+pushd "$BUILD_DIR"
+    zip -r "$ZIPFILE" bugsnag
 popd
-